finally figured out the syntax for shared pointers for p2.cpp

+++ P2/customer.h
@@ -6,7 +6,7 @@
 
 #include "vendor.h"
 
-// Class invarients: 
+// Class invarients:
 
 // Interface invarients:
 
@@ -34,7 +34,7 @@ class Customer
 
     // PRE: N/A
     // POST: N/A
-    double getCurrentBalance();
+    float getCurrentBalance();
 };
 
 #endif
